+++ tests/format_test.py
@@ -1,28 +1,42 @@
-from tempfile import TemporaryDirectory
-from typing import Optional, Iterable
+from typing import Optional, Iterable, Dict
 
 import pytest
 
+from tango.common.testing import TangoTestCase
 from tango.format import _OPEN_FUNCTIONS, DillFormat, JsonFormat
 
 
-@pytest.mark.parametrize("compress", _OPEN_FUNCTIONS.keys())
-def test_iterable_dill_format(compress: Optional[str]):
-    r = (x + 1 for x in range(10))
+class TestFormat(TangoTestCase):
+    @pytest.mark.parametrize("compress", _OPEN_FUNCTIONS.keys())
+    def test_dill_format(self, compress: Optional[str]):
+        artifact = "Hello, World!"
+        format = DillFormat[str](compress)
+        format.write(artifact, self.TEST_DIR)
+        format.checksum(self.TEST_DIR)
+        assert format.read(self.TEST_DIR) == artifact
 
-    with TemporaryDirectory(prefix="test_iterable_dill_format-") as d:
+    @pytest.mark.parametrize("compress", _OPEN_FUNCTIONS.keys())
+    def test_iterable_dill_format(self, compress: Optional[str]):
+        r = (x + 1 for x in range(10))
+
         format = DillFormat[Iterable[int]](compress)
-        format.write(r, d)
-        r2 = format.read(d)
+        format.write(r, self.TEST_DIR)
+        r2 = format.read(self.TEST_DIR)
         assert [x + 1 for x in range(10)] == list(r2)
 
+    @pytest.mark.parametrize("compress", _OPEN_FUNCTIONS.keys())
+    def test_json_format(self, compress: Optional[str]):
+        artifact = {"Hello, World!": "Hi!"}
+        format = JsonFormat[Dict[str, str]](compress)
+        format.write(artifact, self.TEST_DIR)
+        format.checksum(self.TEST_DIR)
+        assert format.read(self.TEST_DIR) == artifact
 
-@pytest.mark.parametrize("compress", _OPEN_FUNCTIONS.keys())
-def test_iterable_json_format(compress: Optional[str]):
-    r = (x + 1 for x in range(10))
+    @pytest.mark.parametrize("compress", _OPEN_FUNCTIONS.keys())
+    def test_iterable_json_format(self, compress: Optional[str]):
+        r = (x + 1 for x in range(10))
 
-    with TemporaryDirectory(prefix="test_iterable_json_format-") as d:
         format = JsonFormat[Iterable[int]](compress)
-        format.write(r, d)
-        r2 = format.read(d)
+        format.write(r, self.TEST_DIR)
+        r2 = format.read(self.TEST_DIR)
         assert [x + 1 for x in range(10)] == list(r2)